requirements.txt
requests~=2.27.1
awscli
flask~=2.1.2
selenium~=4.1.5
Pillow~=9.1.0
names~=0.3.0